+++ packages/typespec-go/test/local/azkeys/custom_client.go
@@ -1,0 +1,20 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+package azkeys
+
+import (
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+)
+
+func NewClient(vaultURL string, options *azcore.ClientOptions) (*Client, error) {
+	client, err := azcore.NewClient(moduleName, moduleVersion, runtime.PipelineOptions{}, options)
+	if err != nil {
+		return nil, err
+	}
+	return &Client{
+		internal:     client,
+		vaultBaseUrl: vaultURL,
+	}, nil
+}