+++ packages/typespec-go/test/local/rawjson/subpkg/zz_rawjson_client.go
@@ -13,11 +13,6 @@ type RawJSONClient struct {
 	endpoint string
 }
 
-// RawJSONClientOptions contains the optional values for creating a [RawJSONClient].
-type RawJSONClientOptions struct {
-	azcore.ClientOptions
-}
-
 // NewRawJSONInputOnlyClient creates a new instance of [RawJSONInputOnlyClient].
 func (client *RawJSONClient) NewRawJSONInputOnlyClient() *RawJSONInputOnlyClient {
 	return &RawJSONInputOnlyClient{